+++ CHANGELOG.md
@@ -12,7 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   fewer `as` casts in their implementation. This caught an issue where
   `EncodeGaugeValue` would not error when encoding some `u64`s that don't fit
   in a `i64`. See [PR 281].
+- Filter out empty metric families, to match the go client. See [PR 279].
 
+[PR 279]: https://github.com/prometheus/client_rust/pull/279
 [PR 281]: https://github.com/prometheus/client_rust/pull/281
 
 ## [0.24.0]